move SDP generation down to the RTP layer.
Hook up RTP/RTCP for STUN. There's a problem where right now, we can get
insane results from STUN (i.e. RTP and RTCP get port-address-translated
to ports that are no longer next to each other, or RTP is now on an odd
port.) This needs to be communicated back up to the phone layer. At the
moment it just spits out an error to stdout.