BYE works again. sigh. Mea maxima stupido.
Cache the local ip address rather than running up a ConnectedDatagramProtocol
each time. In theory, if you've two interfaces, and are calling two different
machines (on each interface), and both of them are behind a firewall (in RFC1918
space), then this could get it wrong. I don't care ;) the performance gain is
worth it.