* undo earlier patch
While raising an exception in this case is convenient if you have an internal error and want to debug, it is inconvenient if you have a "I don't want to answer your call" user action which is encoded into an exception (precisely: a Twisted Failure object).

Too bad those two kinds of things are indistinguishable at this point in the code, but oh well.